+++ features/product-page/ai-context.ts
@@ -1,11 +1,11 @@
 import { createAI } from "ai/rsc";
-import { generateProductDescription } from "./product-description/actions-product-description";
+import { generateRecommendation } from "./recommendation/actions-recommendation";
 
 // Create the AI provider with the initial states and allowed actions
 export const AI = createAI({
   initialAIState: [],
   initialUIState: [],
   actions: {
-    generateImageDescription: generateProductDescription,
+    generateImageDescription: generateRecommendation, // is generateImageDescription used anywhere?
   },
 });
