+++ features/product-page/recommendation/prompt.ts
